+++ sdk/core/azure-core-tracing-opentelemetry/src/test/java/com/azure/core/tracing/opentelemetry/AmqpPropagationFormatUtilTest.java
@@ -4,9 +4,10 @@ package com.azure.core.tracing.opentelemetry;
 
 import com.azure.core.util.Context;
 import com.azure.core.util.tracing.Tracer;
+import io.opentelemetry.api.OpenTelemetry;
 import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.api.trace.SpanContext;
-import io.opentelemetry.api.trace.TracerProvider;
+import io.opentelemetry.sdk.OpenTelemetrySdk;
 import io.opentelemetry.sdk.trace.SdkTracerProvider;
 import org.junit.jupiter.api.Test;
 
@@ -22,9 +23,10 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 @SuppressWarnings("deprecation")
 public class AmqpPropagationFormatUtilTest {
 
-    private static final TracerProvider TRACER_PROVIDER = SdkTracerProvider.builder().build();
+    private static final SdkTracerProvider TRACER_PROVIDER = SdkTracerProvider.builder().build();
+    private static final OpenTelemetry OPEN_TELEMETRY = OpenTelemetrySdk.builder().setTracerProvider(TRACER_PROVIDER).build();
     private static final Tracer TRACER = new OpenTelemetryTracer("test", null, null,
-        new OpenTelemetryTracingOptions().setProvider(TRACER_PROVIDER));
+        new OpenTelemetryTracingOptions().setOpenTelemetry(OPEN_TELEMETRY));
     @Test
     public void extractContextReturnsSpanContext() {