+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/CosmosException.java
@@ -51,7 +51,6 @@ public class CosmosException extends AzureException {
     private static final long serialVersionUID = 1L;
 
     private static final ObjectMapper mapper = new ObjectMapper();
-    private final static String USER_AGENT = Utils.getUserAgent();
 
     /**
      * Status code
@@ -111,7 +110,7 @@ public class CosmosException extends AzureException {
     /**
      * Request URI
      */
-    Uri requestUri;
+    private Uri requestUri;
 
     /**
      * Resource address
@@ -423,12 +422,20 @@ public class CosmosException extends AzureException {
         return Double.parseDouble(value);
     }
 
+    void setRequestUri(Uri requestUri) {
+        this.requestUri = requestUri;
+    }
+
+    Uri getRequestUri() {
+        return this.requestUri;
+    }
+
     @Override
     public String toString() {
         try {
             ObjectNode exceptionMessageNode = mapper.createObjectNode();
             exceptionMessageNode.put("ClassName", getClass().getSimpleName());
-            exceptionMessageNode.put(USER_AGENT_KEY, USER_AGENT);
+            exceptionMessageNode.put(USER_AGENT_KEY, this.getUserAgent());
             exceptionMessageNode.put("statusCode", statusCode);
             exceptionMessageNode.put("resourceAddress", resourceAddress);
             if (cosmosError != null) {
@@ -460,7 +467,7 @@ public class CosmosException extends AzureException {
                 "%s {%s=%s, error=%s, resourceAddress=%s, statusCode=%s, message=%s, causeInfo=%s, responseHeaders=%s, requestHeaders=%s, faultInjectionRuleId=[%s] }",
                 getClass().getSimpleName(),
                 USER_AGENT_KEY,
-                USER_AGENT,
+                this.getUserAgent(),
                 cosmosError,
                 resourceAddress,
                 statusCode,
@@ -568,6 +575,15 @@ public class CosmosException extends AzureException {
         this.sendingRequestHasStarted = hasSendingRequestStarted;
     }
 
+    private String getUserAgent() {
+        String userAgent = Utils.getUserAgent();
+        if (this.requestHeaders != null) {
+            userAgent = this.requestHeaders.getOrDefault(HttpConstants.HttpHeaders.USER_AGENT, userAgent);
+        }
+
+        return userAgent;
+    }
+
     void setFaultInjectionRuleId(String faultInjectionRUleId) {
         this.faultInjectionRuleId = faultInjectionRUleId;
     }
@@ -636,6 +652,16 @@ public class CosmosException extends AzureException {
                     @Override
                     public List<String> getFaultInjectionEvaluationResults(CosmosException cosmosException) {
                         return cosmosException.getFaultInjectionEvaluationResults();
+                    }
+
+                    @Override
+                    public void setRequestUri(CosmosException cosmosException, Uri requestUri) {
+                        cosmosException.setRequestUri(requestUri);
+                    }
+
+                    @Override
+                    public Uri getRequestUri(CosmosException cosmosException) {
+                        return cosmosException.getRequestUri();
                     }
                 });
     }