+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/ImplementationBridgeHelpers.java
@@ -18,6 +18,7 @@ import com.azure.cosmos.CosmosDiagnostics;
 import com.azure.cosmos.CosmosDiagnosticsContext;
 import com.azure.cosmos.CosmosDiagnosticsHandler;
 import com.azure.cosmos.CosmosDiagnosticsThresholds;
+import com.azure.cosmos.CosmosEndToEndOperationLatencyPolicyConfig;
 import com.azure.cosmos.CosmosException;
 import com.azure.cosmos.DirectConnectionConfig;
 import com.azure.cosmos.GlobalThroughputControlConfig;
@@ -28,6 +29,7 @@ import com.azure.cosmos.implementation.clienttelemetry.ClientTelemetry;
 import com.azure.cosmos.implementation.clienttelemetry.CosmosMeterOptions;
 import com.azure.cosmos.implementation.clienttelemetry.MetricCategory;
 import com.azure.cosmos.implementation.clienttelemetry.TagName;
+import com.azure.cosmos.implementation.directconnectivity.Uri;
 import com.azure.cosmos.implementation.directconnectivity.ContainerDirectConnectionMetadata;
 import com.azure.cosmos.implementation.directconnectivity.rntbd.RntbdChannelStatistics;
 import com.azure.cosmos.implementation.faultinjection.IFaultInjectorProvider;
@@ -54,6 +56,7 @@ import com.azure.cosmos.models.CosmosQueryRequestOptions;
 import com.azure.cosmos.models.FeedResponse;
 import com.azure.cosmos.models.ModelBridgeInternal;
 import com.azure.cosmos.models.PartitionKey;
+import com.azure.cosmos.models.PriorityLevel;
 import com.azure.cosmos.util.CosmosPagedFlux;
 import com.azure.cosmos.util.UtilBridgeInternal;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -213,9 +216,6 @@ public class ImplementationBridgeHelpers {
             DirectConnectionConfig setHealthCheckTimeoutDetectionEnabled(
                 DirectConnectionConfig directConnectionConfig, boolean timeoutDetectionEnabled);
             boolean isHealthCheckTimeoutDetectionEnabled(DirectConnectionConfig directConnectionConfig);
-            DirectConnectionConfig setMinConnectionPoolSizePerEndpoint(
-                    DirectConnectionConfig directConnectionConfig, int minConnectionPoolSizePerEndpoint);
-            int getMinConnectionPoolSizePerEndpoint(DirectConnectionConfig directConnectionConfig);
         }
     }
 
@@ -267,6 +267,7 @@ public class ImplementationBridgeHelpers {
             RequestOptions toRequestOptions(CosmosQueryRequestOptions queryRequestOptions);
             CosmosDiagnosticsThresholds getDiagnosticsThresholds(CosmosQueryRequestOptions options);
             void applyMaxItemCount(CosmosQueryRequestOptions requestOptions, CosmosPagedFluxOptions fluxOptions);
+            CosmosEndToEndOperationLatencyPolicyConfig getEndToEndOperationLatencyPolicyConfig(CosmosQueryRequestOptions options);
         }
     }
 
@@ -712,6 +713,10 @@ public class ImplementationBridgeHelpers {
             void addClientSideDiagnosticsToFeed(
                 CosmosDiagnostics cosmosDiagnostics,
                 Collection<ClientSideRequestStatistics> requestStatistics);
+
+            void setSamplingRateSnapshot(CosmosDiagnostics cosmosDiagnostics, double samplingRate);
+
+            CosmosDiagnostics create(DiagnosticsClientContext clientContext, double samplingRate);
         }
     }
 
@@ -800,6 +805,8 @@ public class ImplementationBridgeHelpers {
             Collection<ClientSideRequestStatistics> getDistinctCombinedClientSideRequestStatistics(CosmosDiagnosticsContext ctx);
 
             String getSpanName(CosmosDiagnosticsContext ctx);
+
+            void setSamplingRateSnapshot(CosmosDiagnosticsContext ctx, double samplingRate);
         }
     }
 
@@ -1295,6 +1302,8 @@ public class ImplementationBridgeHelpers {
 
             void setFaultInjectionEvaluationResults(CosmosException cosmosException, List<String> faultInjectionRuleEvaluationResults);
             List<String> getFaultInjectionEvaluationResults(CosmosException cosmosException);
+            void setRequestUri(CosmosException cosmosException, Uri requestUri);
+            Uri getRequestUri(CosmosException cosmosException);
         }
     }
 
@@ -1363,8 +1372,47 @@ public class ImplementationBridgeHelpers {
             boolean isTransportLevelTracingEnabled(CosmosClientTelemetryConfig config);
             Tracer getOrCreateTracer(CosmosClientTelemetryConfig config);
             void setUseLegacyTracing(CosmosClientTelemetryConfig config, boolean useLegacyTracing);
-
             void setTracer(CosmosClientTelemetryConfig config, Tracer tracer);
+            double getSamplingRate(CosmosClientTelemetryConfig config);
+        }
+    }
+
+    public static final class PriorityLevelHelper {
+        private final static AtomicBoolean priorityLevelClassLoaded = new AtomicBoolean(false);
+        private final static AtomicReference<PriorityLevelAccessor> accessor = new AtomicReference<>();
+
+        private PriorityLevelHelper() {
+        }
+
+        public static PriorityLevelAccessor getPriorityLevelAccessor() {
+            if (!priorityLevelClassLoaded.get()) {
+                logger.debug("Initializing PriorityLevelAccessor...");
+                initializeAllAccessors();
+            }
+
+            PriorityLevelAccessor snapshot = accessor.get();
+            if (snapshot == null) {
+                logger.error("PriorityLevelAccessor is not initialized yet!");
+                System.exit(9728); // Using a unique status code here to help debug the issue.
+            }
+
+            return snapshot;
+        }
+
+        public static void setPriorityLevelAccessor(final PriorityLevelAccessor newAccessor) {
+
+            assert(newAccessor != null);
+
+            if (!accessor.compareAndSet(null, newAccessor)) {
+                logger.debug("PriorityLevelAccessor already initialized!");
+            } else {
+                logger.debug("Setting PriorityLevelAccessor...");
+                priorityLevelClassLoaded.set(true);
+            }
+        }
+
+        public interface PriorityLevelAccessor {
+            byte getPriorityValue(PriorityLevel level);
         }
     }