+++ sdk/core/azure-core-tracing-opentelemetry/CHANGELOG.md
@@ -1,14 +1,25 @@
 # Release History
 
-## 1.0.0-beta.35 (Unreleased)
+## 1.0.0-beta.36 (Unreleased)
 
 ### Features Added
 
 ### Breaking Changes
 
+- Replaced `OpenTelemetryTracingOptions.setProvider` with `OpenTelemetryTracingOptions.setOpenTelemetry` method. Instead of `io.opentelemetry.api.trace.TracerProvider` instance
+  it now takes `io.opentelemetry.api.OpenTelemetry` container.
+
 ### Bugs Fixed
 
 ### Other Changes
+
+## 1.0.0-beta.35 (2023-05-04)
+
+### Other Changes
+
+#### Dependency Updates
+
+- Upgraded `azure-core` from `1.38.0` to `1.39.0`.
 
 ## 1.0.0-beta.34 (2023-04-07)